Conexión IMAP correcta

+++ mondialRelayWP.php
@@ -13,7 +13,7 @@
  * Plugin URI:        https://github.com/PedroGarcia-BWH/MondialRelayYoreleoWP
  * Description:       Conexión con la API de Mondial Relay para la gestión de envíos y datos de los clientes de Yoreleo
  * Version:           alpha-0.0.5
- * Requires at least: 8.1
+ * Requires at least: 5.2
  * Requires PHP:      8.1
  * Author:            Pedro José García Romera
  * Author URI:        https://github.com/PedroGarcia-BWH
@@ -21,7 +21,6 @@
  * License URI:       http://www.gnu.org/licenses/gpl-2.0.txt
  * Text Domain:       yoreleo-mondial-relay
  */
-
 
 if (!defined('WPINC')){
 	die;
@@ -36,6 +35,11 @@ function my_plugin_activation() {
     $folder_path = $theme_dir . '/ultimate-member/' . $new_folder;
     $file_path = $folder_path . '/order-popup.php';
 
+    if (!extension_loaded('imap')) {
+        echo 'La extensión IMAP no está habilitada';
+        exit();
+    }
+
     if (file_exists($folder_path)) {
         if(file_exists($file_path)) unlink($file_path);
     } else {
@@ -45,6 +49,20 @@ function my_plugin_activation() {
     $file_content = file_get_contents(plugin_dir_path(__FILE__)."public/templates/order-popup.php");
 
     file_put_contents($file_path, $file_content);
+
+}
+
+
+add_action( 'upgrader_process_complete', 'my_plugin_upgrade_action', 10, 2 );
+
+function my_plugin_upgrade_action( $upgrader_object, $options ) {
+    if ( $options['type'] == 'plugin' && isset( $options['plugins'] ) ) {
+        foreach ( $options['plugins'] as $plugin ) {
+            if ( $plugin == 'MondialRelayYoreleoWP/mondialRelayWP.php' ) { // Reemplaza "my-plugin/my-plugin.php" por la ruta del plugin que quieras comprobar
+                my_plugin_activation();
+            }
+        }
+    }
 }
 
 
@@ -61,3 +79,24 @@ function my_custom_styles() {
 add_action( 'wp_enqueue_scripts', 'my_custom_styles' );
 
 
+function my_account_page_default_tabs( $tabs ) {
+    //$tabs['none'] = __( 'None', 'ultimate-member' );
+    //unset( $tabs['orders'] ); // Remueve la opción de órdenes
+    $tabs['default_tab'] = ''; // Cambia el nombre de la pestaña de órdenes
+    print_r($tabs);
+    return $tabs;
+}
+add_filter( 'um_account_page_default_tabs_hook', 'my_account_page_default_tabs', 999, 1 );
+
+
+/*wp_schedule_event( time(), '5_minutes', 'check_email' );
+
+// Agregar un nuevo intervalo de tiempo de 5 minutos
+add_filter( 'cron_schedules', function( $schedules ) {
+    $schedules['5_minutes'] = array(
+        'interval' => 300,
+        'display'  => esc_html__( 'Every 5 Minutes', 'textdomain' ),
+    );
+    return $schedules;
+} );
+*/
